+++ previous-versions/files/RELEASES.md
@@ -19,3 +19,13 @@ Based on:
 - [python v0.2.0] previous-versions/files
 ### Releases
 - [PyPI v0.2.0] https://pypi.org/project/codat-files/0.2.0 - previous-versions/files
+
+## 2023-10-30 01:24:10
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Files.yaml
+- Speakeasy CLI 1.109.0 (2.173.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v0.2.1] previous-versions/files
+### Releases
+- [PyPI v0.2.1] https://pypi.org/project/codat-files/0.2.1 - previous-versions/files